Adjust to the CGFloat change for NSFont and NSAffineTransform done in base and gui.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/back/trunk@29689 72102866-910b-0410-8b05-ffd578937521

+++ Headers/cairo/CairoFontInfo.h
@@ -38,7 +38,7 @@
 	cairo_scaled_font_t *_scaled;
 	CairoFaceInfo *_faceInfo;
 	BOOL _screenFont;
-	float lineHeight;
+	CGFloat lineHeight;
 
 	unsigned int _cacheSize;
 	unsigned int *_cachedGlyphs;
